Create securePassword
In this code, we use the random module to generate a secure and random password. The string module provides different sets of characters that can be used in the password generation process.

The generate_password() function takes an optional parameter length which specifies the length of the generated password. By default, it generates a 12-character long password.